type and interface transferred to the central folder
